% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-gephi.R
\name{gephi_palettes}
\alias{gephi_palettes}
\title{Gephi generative palette names}
\usage{
gephi_palettes()
}
\value{
Character vector of Gephi generative palette identifiers.
}
\description{
Gephi generative palette names
}
\examples{
gephi_palettes()
}
