% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-iterm.R
\name{iterm_palettes}
\alias{iterm_palettes}
\title{iTerm color palette names}
\usage{
iterm_palettes()
}
\value{
Character vector of palette names.
}
\description{
iTerm color palette names
}
\examples{
iterm_palettes()
}
