% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-cosmic.R
\name{pal_cosmic}
\alias{pal_cosmic}
\title{COSMIC Color Palettes}
\usage{
pal_cosmic(palette = c("hallmarks"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is two available options: \code{"signature_substitutions"}
(6-color palette) and \code{"hallmarks"} (10-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Color palettes inspired by the colors used projects from the
\href{https://cancer.sanger.ac.uk/cosmic}{Catalogue Of
Somatic Mutations in Cancers (COSMIC)}
}
\examples{
library("scales")
show_col(pal_cosmic("hallmark")(10))
show_col(pal_cosmic("hallmark", alpha = 0.6)(10))
show_col(pal_cosmic("signature_substitutions")(6))
show_col(pal_cosmic("signature_substitutions", alpha = 0.6)(6))
}
\author{
Nan Xiao <\email{me@nanx.me}> |
<\href{https://nanx.me}{https://nanx.me}>

Joshua H. Cook <\email{joshuacook0023@gmail.com}> |
<\href{https://github.com/jhrcook}{GitHub/jhrcook}>
}
