% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-frontiers.R
\name{pal_frontiers}
\alias{pal_frontiers}
\title{Frontiers Color Palettes}
\usage{
pal_frontiers(palette = c("default"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"}
(10-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Color palettes inspired by the colors used in \emph{Frontiers}.
}
\examples{
library("scales")
show_col(pal_frontiers("default")(7))
show_col(pal_frontiers("default", alpha = 0.6)(7))
}
\author{
Clara Jégousse <\email{cat3@hi.is}>
}
