% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-futurama.R
\name{pal_futurama}
\alias{pal_futurama}
\title{The Futurama Color Palettes}
\usage{
pal_futurama(palette = c("planetexpress"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"planetexpress"}
(12-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Color palettes inspired by the colors used in \emph{Futurama}.
}
\examples{
library("scales")
show_col(pal_futurama("planetexpress")(12))
show_col(pal_futurama("planetexpress", alpha = 0.6)(12))
}
\author{
Nan Xiao <\email{me@nanx.me}> |
<\href{https://nanx.me}{https://nanx.me}>
}
