% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-jco.R
\name{pal_jco}
\alias{pal_jco}
\title{Journal of Clinical Oncology Color Palettes}
\usage{
pal_jco(palette = c("default"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"}
(10-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Color palette inspired by plots in \emph{Journal of Clinical Oncology}.
}
\examples{
library("scales")
show_col(pal_jco("default")(10))
show_col(pal_jco("default", alpha = 0.6)(10))
}
\author{
Nan Xiao <\email{me@nanx.me}> |
<\href{https://nanx.me}{https://nanx.me}>
}
