% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-observable.R
\name{pal_observable}
\alias{pal_observable}
\title{Observable 10 color palette}
\usage{
pal_observable(palette = c("observable10"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"observable10"}
(10-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}
}
\description{
The Observable 10 palette.
}
\examples{
library("scales")
show_col(pal_observable("observable10")(10))
show_col(pal_observable("observable10", alpha = 0.6)(10))
}
\references{
Pettiross J (2023). "Crafting data colors and staying on brand."
\emph{Observable blog}. \url{https://observablehq.com/blog/crafting-data-colors}
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
