% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-primer.R
\name{pal_primer}
\alias{pal_primer}
\title{Primer design system palette}
\usage{
pal_primer(palette = c("mark17"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"mark17"}
(17-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}
}
\description{
The Primer design system data visualization palette.
}
\examples{
library("scales")
show_col(pal_primer("mark17")(17))
show_col(pal_primer("mark17", alpha = 0.6)(17))
}
\references{
GitHub (2025). "Primer data visualization colors."
\url{https://primer.style/product/ui-patterns/data-visualization/}
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
