% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-squidgame.R
\name{pal_squidgame}
\alias{pal_squidgame}
\title{Squid Game Color Palettes}
\usage{
pal_squidgame(palette = c("stair_room"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"stair_room"}
(8-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Color palettes inspired by the colors used in \emph{Squid Game}.
}
\examples{
library("scales")
show_col(pal_squidgame("stair_room")(8))
show_col(pal_squidgame("stair_room", alpha = 0.6)(8))
}
\author{
Keith Turner <\email{khturner@gmail.com}> |
<\href{https://amazon.com}{https://amazon.com}>
}
