% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-ucscgb.R
\name{pal_ucscgb}
\alias{pal_ucscgb}
\title{UCSC Genome Browser Color Palette}
\usage{
pal_ucscgb(palette = c("default"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"}
(26-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Color palette from UCSC Genome Browser chromosome colors.
}
\examples{
library("scales")
show_col(pal_ucscgb("default")(26))
show_col(pal_ucscgb("default", alpha = 0.6)(26))
}
\author{
Nan Xiao <\email{me@nanx.me}> |
<\href{https://nanx.me}{https://nanx.me}>
}
