% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsci.R
\name{palette.aaas}
\alias{palette.aaas}
\title{Color Palette for AAAS Journals}
\usage{
palette.aaas()
}
\description{
A 10-color discrete color palette based on the colors
used in figures in AAAS journals such as Science.
}

