% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsci.R
\name{palette.jco}
\alias{palette.jco}
\title{Color Palette for Journal of Clinical Oncology (JCO)}
\usage{
palette.jco()
}
\description{
A 10-color discrete color palette based on the colors
used in figures in Journal of Clinical Oncology.
}

