% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsci.R
\name{palette.lancet}
\alias{palette.lancet}
\title{Color Palette for Lancet Journals}
\usage{
palette.lancet()
}
\description{
A 10-color discrete color palette based on the colors
used in figures in Lancet Oncology.
}

