% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsci.R
\name{palette.npg}
\alias{palette.npg}
\title{Color Palette for NPG Journals}
\usage{
palette.npg()
}
\description{
A 10-color discrete color palette based on the colors
used in figures in NPG journals such as Nature Reviews Cancer.
}

