% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous-vaccines.R
\name{rgb_vaccine}
\alias{rgb_vaccine}
\title{The Vaccine Color Palette}
\usage{
rgb_vaccine(palette = c("default"), n = 11, alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"}
(continuous palette with 11 base colors).}

\item{n}{Number of individual colors to be generated.}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{reverse}{Logical. Should the order of the colors be reversed?}
}
\description{
Color palette inspired by the colors used by the heatmaps
created by Tynan DeBold and Dov Friedman in the article
"Battling Infectious Diseases in the 20th Century: The Impact of Vaccines".
}
\note{
Since this color palette is biased to one tail of the
spectrum, it is more suitable for displaying heavy tail data.
For example, if a lot of values are concentrated around zero
in a set of numbers ranging from 0 to 1.
In this case, the biased palette can help us better distinguish
the difference between the values in the far end of 0.
}
\examples{
library("scales")
show_col(pal_vaccine("default")(11))
show_col(pal_vaccine("default", n = 30, alpha = 0.6, reverse = TRUE)(30))
}
\author{
Nan Xiao <\email{me@nanx.me}> |
<\href{https://nanx.me}{https://nanx.me}>
}
\references{
\href{http://graphics.wsj.com/infectious-diseases-and-vaccines/}{Battling Infectious Diseases in the 20th Century: The Impact of Vaccines}.
}

