% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-aaas.R
\name{scale_color_aaas}
\alias{scale_color_aaas}
\alias{scale_colour_aaas}
\alias{scale_fill_aaas}
\title{AAAS journal color scales}
\usage{
scale_color_aaas(palette = c("default"), alpha = 1, ...)

scale_colour_aaas(palette = c("default"), alpha = 1, ...)

scale_fill_aaas(palette = c("default"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"}
(10-color palette inspired by \emph{Science}).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_aaas]{pal_aaas()}} for details.
}
\examples{
example_scatterplot() + scale_color_aaas()
example_barplot() + scale_fill_aaas()
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
