% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-atlassian.R
\name{scale_color_atlassian}
\alias{scale_color_atlassian}
\alias{scale_colour_atlassian}
\alias{scale_fill_atlassian}
\title{Atlassian Design System color scales}
\usage{
scale_color_atlassian(palette = c("categorical8"), alpha = 1, ...)

scale_colour_atlassian(palette = c("categorical8"), alpha = 1, ...)

scale_fill_atlassian(palette = c("categorical8"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"categorical8"}
(8-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_atlassian]{pal_atlassian()}} for details.
}
\examples{
example_scatterplot() + scale_color_atlassian()
example_barplot() + scale_fill_atlassian()
}
\references{
Atlassian (2025). "Atlassian Design System categorical chart colors."
\url{https://atlassian.design/foundations/color-new/data-visualization-color}
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
