% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-futurama.R
\name{scale_color_futurama}
\alias{scale_color_futurama}
\alias{scale_colour_futurama}
\alias{scale_fill_futurama}
\title{Futurama color scales}
\usage{
scale_color_futurama(palette = c("planetexpress"), alpha = 1, ...)

scale_colour_futurama(palette = c("planetexpress"), alpha = 1, ...)

scale_fill_futurama(palette = c("planetexpress"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"planetexpress"}
(12-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_futurama]{pal_futurama()}} for details.
}
\examples{
example_scatterplot() + scale_color_futurama()
example_barplot() + scale_fill_futurama()
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
