% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-gephi.R
\name{scale_color_gephi}
\alias{scale_color_gephi}
\alias{scale_colour_gephi}
\alias{scale_fill_gephi}
\title{Gephi color scales}
\usage{
scale_color_gephi(palette = gephi_palettes(), alpha = 1, ...)

scale_colour_gephi(palette = gephi_palettes(), alpha = 1, ...)

scale_fill_gephi(palette = gephi_palettes(), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type. See \code{\link[=gephi_palettes]{gephi_palettes()}} for available options.}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_gephi]{pal_gephi()}} for details.
}
\examples{
set.seed(42)
example_scatterplot() + scale_color_gephi()
example_barplot() + scale_fill_gephi("fancy_light")
}
