% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-jama.R
\name{scale_color_jama}
\alias{scale_color_jama}
\alias{scale_colour_jama}
\alias{scale_fill_jama}
\title{Journal of the American Medical Association color scales}
\usage{
scale_color_jama(palette = c("default"), alpha = 1, ...)

scale_colour_jama(palette = c("default"), alpha = 1, ...)

scale_fill_jama(palette = c("default"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"}
(7-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_jama]{pal_jama()}} for details.
}
\examples{
example_scatterplot() + scale_color_jama()
example_barplot() + scale_fill_jama()
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
