% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-jco.R
\name{scale_color_jco}
\alias{scale_color_jco}
\alias{scale_colour_jco}
\alias{scale_fill_jco}
\title{Journal of Clinical Oncology color scales}
\usage{
scale_color_jco(palette = c("default"), alpha = 1, ...)

scale_colour_jco(palette = c("default"), alpha = 1, ...)

scale_fill_jco(palette = c("default"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"}
(10-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_jco]{pal_jco()}} for details.
}
\examples{
example_scatterplot() + scale_color_jco()
example_barplot() + scale_fill_jco()
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
