% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-locuszoom.R
\name{scale_color_locuszoom}
\alias{scale_color_locuszoom}
\alias{scale_colour_locuszoom}
\alias{scale_fill_locuszoom}
\title{LocusZoom color scales}
\usage{
scale_color_locuszoom(palette = c("default"), alpha = 1, ...)

scale_colour_locuszoom(palette = c("default"), alpha = 1, ...)

scale_fill_locuszoom(palette = c("default"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"}
(7-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_locuszoom]{pal_locuszoom()}} for details.
}
\examples{
example_scatterplot() + scale_color_locuszoom()
example_barplot() + scale_fill_locuszoom()
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
