% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-npg.R
\name{scale_color_npg}
\alias{scale_color_npg}
\alias{scale_colour_npg}
\alias{scale_fill_npg}
\title{NPG journal color scales}
\usage{
scale_color_npg(palette = c("nrc"), alpha = 1, ...)

scale_colour_npg(palette = c("nrc"), alpha = 1, ...)

scale_fill_npg(palette = c("nrc"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"nrc"}
(10-color palette inspired by \emph{Nature Reviews Cancer}).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_npg]{pal_npg()}} for details.
}
\examples{
example_scatterplot() + scale_color_npg()
example_barplot() + scale_fill_npg()
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
