% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-primer.R
\name{scale_color_primer}
\alias{scale_color_primer}
\alias{scale_colour_primer}
\alias{scale_fill_primer}
\title{Primer color scales}
\usage{
scale_color_primer(palette = c("mark17"), alpha = 1, ...)

scale_colour_primer(palette = c("mark17"), alpha = 1, ...)

scale_fill_primer(palette = c("mark17"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"mark17"}
(17-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_primer]{pal_primer()}} for details.
}
\examples{
example_scatterplot() + scale_color_primer()
example_barplot() + scale_fill_primer()
}
\references{
GitHub (2025). "Primer data visualization colors."
\url{https://primer.style/product/ui-patterns/data-visualization/}
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
