% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-rickandmorty.R
\name{scale_color_rickandmorty}
\alias{scale_color_rickandmorty}
\alias{scale_colour_rickandmorty}
\alias{scale_fill_rickandmorty}
\title{Rick and Morty color scales}
\usage{
scale_color_rickandmorty(palette = c("schwifty"), alpha = 1, ...)

scale_colour_rickandmorty(palette = c("schwifty"), alpha = 1, ...)

scale_fill_rickandmorty(palette = c("schwifty"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"schwifty"}
(12-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_rickandmorty]{pal_rickandmorty()}} for details.
}
\examples{
example_scatterplot() + scale_color_rickandmorty()
example_barplot() + scale_fill_rickandmorty()
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
