% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-simpsons.R
\name{scale_color_simpsons}
\alias{scale_color_simpsons}
\alias{scale_colour_simpsons}
\alias{scale_fill_simpsons}
\title{The Simpsons color scales}
\usage{
scale_color_simpsons(palette = c("springfield"), alpha = 1, ...)

scale_colour_simpsons(palette = c("springfield"), alpha = 1, ...)

scale_fill_simpsons(palette = c("springfield"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"springfield"}
(16-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_simpsons]{pal_simpsons()}} for details.
}
\examples{
example_scatterplot() + scale_color_simpsons()
example_barplot() + scale_fill_simpsons()
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
