% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-startrek.R
\name{scale_color_startrek}
\alias{scale_color_startrek}
\alias{scale_colour_startrek}
\alias{scale_fill_startrek}
\title{Star Trek color scales}
\usage{
scale_color_startrek(palette = c("uniform"), alpha = 1, ...)

scale_colour_startrek(palette = c("uniform"), alpha = 1, ...)

scale_fill_startrek(palette = c("uniform"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"uniform"}
(7-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_startrek]{pal_startrek()}} for details.
}
\examples{
example_scatterplot() + scale_color_startrek()
example_barplot() + scale_fill_startrek()
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
