% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-tron.R
\name{scale_color_tron}
\alias{scale_color_tron}
\alias{scale_colour_tron}
\alias{scale_fill_tron}
\title{Tron Legacy color scales}
\usage{
scale_color_tron(palette = c("legacy"), alpha = 1, ...)

scale_colour_tron(palette = c("legacy"), alpha = 1, ...)

scale_fill_tron(palette = c("legacy"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"legacy"}
(7-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_tron]{pal_tron()}} for details.
}
\examples{
example_scatterplot() +
  ggplot2::theme_dark() +
  ggplot2::theme(
    panel.background = ggplot2::element_rect(fill = "#2D2D2D"),
    legend.key = ggplot2::element_rect(fill = "#2D2D2D")
  ) +
  scale_color_tron()

example_barplot() +
  ggplot2::theme_dark() +
  ggplot2::theme(
    panel.background = ggplot2::element_rect(fill = "#2D2D2D")
  ) +
  scale_fill_tron()
}
\author{
Nan Xiao | \email{me@nanx.me} | \url{https://nanx.me}
}
