% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-blood.R
\name{get_ash}
\alias{get_ash}
\title{Get a named vector of colours}
\usage{
get_ash(palette = c("blood", "lymphgen", "coo"))
}
\arguments{
\item{palette}{}
}
\value{

}
\description{
Get a named vector of colours
}
