% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsci-package.R
\docType{package}
\name{ggsci-package}
\alias{ggsci-package}
\title{Scientific Journal and Sci-Fi Themed Color Palettes for 'ggplot2'}
\description{
Scientific Journal and Sci-Fi Themed Color Palettes for 'ggplot2'
}
\details{
Opened the vignette via \code{vignette("ggsci")}.

\tabular{ll}{ Package: \tab ggsci\cr
Type: \tab Package\cr
License: \tab GPL\cr }
}
\author{
Nan Xiao <\email{nanx@uchicago.edu}>
        Miaozhu Li <\email{miaozhu.li@duke.edu}>
}

