% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-blood.R
\name{pal_ash}
\alias{pal_ash}
\title{ASH Journal Color Palettes}
\usage{
pal_ash(palette = c("blood", "lymphgen"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"blood"}
(15-color palette used by the American Society of Hematology's \emph{Blood Journal}).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Color palettes inspired by plots in journals published by
ASH such as \emph{Blood}.
}
\examples{
library("scales")
show_col(pal_ash("blood")(15))
show_col(pal_ash("blood", alpha = 0.6)(15))
}
\author{
Ryan Morin <\email{rdmorin@sfu.ca}> |
}
