% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-bmj.R
\name{scale_color_bmj}
\alias{scale_color_bmj}
\alias{scale_colour_bmj}
\alias{scale_fill_bmj}
\title{BMJ color scales}
\usage{
scale_color_bmj(palette = c("default"), alpha = 1, ...)

scale_colour_bmj(palette = c("default"), alpha = 1, ...)

scale_fill_bmj(palette = c("default"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is one available option: \code{"default"}
(9-color palette).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices:rgb]{grDevices::rgb()}} for details.}

\item{...}{Additional parameters for \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
See \code{\link[=pal_bmj]{pal_bmj()}} for details.
}
\examples{
example_scatterplot() + scale_color_bmj()
example_barplot() + scale_fill_bmj()
}
\references{
\url{https://technology.bmj.com/living-style-guide/colour.html}
}
\author{
Hui Chen | \email{huichen@zju.edu.cn}
}
